Refactoring to leave the file-saving to the editor/plugin  ... not finished!
Need to do this so that we can support publishing to FTP/SFTP/FTPS (and SSH)

I know this can be done with a single ftp or scp command, witch was
originally intended in cmsJZ, but with it built-in with the editor
it needs to be platfom independent.

Unix/Linux vs Windows will need a blog post about their different philosofies.