Had a long discussion on IRC about NPM and nodejs packages and dependencies.
The most common workflow is to tread the node_modules folder as a black box
and leave it outside source control management.

For this project though, to decouple plugins, and treat all code, even
dependencies as part of the editor, each plugin should have it's own
node_modules folder if it uses any nodejs modules.

Only modules that are used by the core should be located in the
root node_modules folder.